Fixed JVM crash due to EXCEPTION_ACCESS_VIOLATION
"The function get_java_var_long() in SerialImp.c casts the Java-long-value to a long in case of type == 'J'. But Java-Long is 64 bit and C-long is only 32 bit. Later this long is cast to a 64-bit-memory-address -> crash. A possible solution is to return always size_t instead of long."

https://github.com/NeuronRobotics/nrjavaserial/issues/131
https://github.com/NeuronRobotics/nrjavaserial/pull/133/
https://github.com/NeuronRobotics/nrjavaserial/pull/135/